+++ smart_trav_plan/SmartTrav/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 class SmartTravConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'smart_trav_plan.SmartTrav'
+    name = 'SmartTrav'
