+++ smart_trav_plan/SmartTrav/static/SmartTrav/css/js/script.js
