+++ smart_trav_plan/SmartTrav/static/SmartTrav/css/style.css
