+++ smart_trav_plan/SmartTrav/static/images/mactan.jpg
