+++ smart_trav_plan/SmartTrav/static/images/magellan.jpg
