+++ smart_trav_plan/SmartTrav/templates/SmartTrav/accounts/about.html
@@ -348,14 +348,14 @@
             </a>
             <div class="nav-links">
                 <a href="/">Home</a>
-                <a href="/accounts/about/" class="btn-about">About</a>
-                <a href="/accounts/service" class="btn-service">Service</a>
+                <a href="{% url 'about' %}">About</a>
+                <a href="/service/" class="btn-service">Service</a>
                 <a href="#">Pricing</a>
                 <a href="#">Contact</a>
             </div>
             <div class="auth-buttons">
-                <a href="/accounts/login/" class="btn-login">Log in</a>
-                <a href="/accounts/signup/" class="btn-signup">Sign up</a>
+                <a href="/SmartTrav/login/" class="btn-login">Log in</a>
+                <a href="/SmartTrav/signup/" class="btn-signup">Sign up</a>
             </div>
         </div>
     </nav>
@@ -474,7 +474,7 @@
         <div class="cta-section">
             <h2>Ready to Start Your Adventure?</h2>
             <p>Join thousands of travelers who trust SmartTrav for their Mactan journey</p>
-            <a href="/accounts/signup/" class="btn-cta">Get Started Today</a>
+            <a href="/SmartTrav/signup/" class="btn-cta">Get Started Today</a>
         </div>
     </div>
