+++ smart_trav_plan/SmartTrav/templates/SmartTrav/accounts/dashboard.html
