+++ smart_trav_plan/SmartTrav/templates/SmartTrav/accounts/service.html
@@ -190,8 +190,8 @@
         <div class="container-fluid">
             <div class="navbar-brand">✈️ SmartTrav</div>
             <div class="nav-links d-none d-md-flex">
-                <a href="/accounts/about/" class="btn-about">About</a>
-                <a href="/accounts/service" class="btn-service">Service</a>
+                <a href="/SmartTrav/about/" class="btn-about">About</a>
+                <a href="/SmartTrav/service" class="btn-service">Service</a>
                 <a href="#">Pricing</a>
                 <a href="#">Contact</a>
             </div>
