+++ smart_trav_plan/SmartTrav/templates/SmartTrav/accounts/signup.html
@@ -291,8 +291,8 @@
         <div class="container-fluid">
             <div class="navbar-brand">✈️ SmartTrav</div>
             <div class="nav-links d-none d-md-flex">
-                <a href="/accounts/about/" class="btn-about">About</a>
-                <a href="/accounts/service" class="btn-service">Service</a>
+                <a href="/SmartTrav/about/" class="btn-about">About</a>
+                <a href="/SmartTrav/service" class="btn-service">Service</a>
                 <a href="#pricing">Pricing</a>
                 <a href="#contact">Contact</a>
             </div>
