+++ smart_trav_plan/SmartTrav/templates/base.html
