--- smart_trav_plan/accounts/__init__.py
@@ -1,1 +1,0 @@
-# accounts/__init__.py
