--- smart_trav_plan/accounts/admin.py
@@ -1,5 +1,0 @@
-# accounts/admin.py
-from django.contrib import admin
-
-# Register your models here.
-# Using Django's built-in User model, no custom models to register
