--- smart_trav_plan/accounts/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-class AccountsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'smart_trav_plan.accounts'
