--- smart_trav_plan/accounts/models.py
@@ -1,5 +1,0 @@
-# accounts/models.py
-from django.db import models
-
-# Using Django's built-in User model for authentication
-# No custom models needed for basic authentication
