+++ smart_trav_plan/manage.py
