+++ smart_trav_plan/smart_trav_plan/settings.py
@@ -2,15 +2,23 @@ from pathlib import Path
 import os
 from dotenv import load_dotenv
 import dj_database_url
+from django.contrib.messages import constants as messages
+
+# Build paths inside the project like this: BASE_DIR / 'subdir'.
+BASE_DIR = Path(__file__).resolve().parent.parent
 
 # Load environment variables from .env
-BASE_DIR = Path(__file__).resolve().parent.parent
 load_dotenv(BASE_DIR / ".env")
 
+# SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = os.environ.get("SECRET_KEY", "replace-this-with-a-secure-key")
+
+# SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = os.environ.get("DEBUG", "True") == "True"
+
 ALLOWED_HOSTS = ["127.0.0.1", "localhost"]
 
+# Application definition
 INSTALLED_APPS = [
     'django.contrib.admin',
     'django.contrib.auth',
@@ -20,8 +28,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
 
     # custom apps
-    'smart_trav_plan.accounts',
-    'smart_trav_plan.SmartTrav',
+    'SmartTrav',  # ✅ Keep only this one
 ]
 
 MIDDLEWARE = [
@@ -39,7 +46,7 @@ ROOT_URLCONF = 'smart_trav_plan.urls'
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [BASE_DIR / "smart_trav_plan" / "templates"],  # ✅ root templates folder
+        'DIRS': [BASE_DIR / "smart_trav_plan" / "templates"],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
@@ -54,16 +61,7 @@ TEMPLATES = [
 
 WSGI_APPLICATION = 'smart_trav_plan.wsgi.application'
 
-# ✅ Database configuration from .env (Supabase Session Pooler)
-import os
-import dj_database_url
-from dotenv import load_dotenv
-from pathlib import Path
-
-BASE_DIR = Path(__file__).resolve().parent.parent
-
-load_dotenv(BASE_DIR / ".env")
-
+# Database - Supabase Session Pooler
 DATABASES = {
     'default': dj_database_url.config(
         default=os.environ.get("DATABASE_URL"),
@@ -72,8 +70,7 @@ DATABASES = {
     )
 }
 
-
-
+# Password validation
 AUTH_PASSWORD_VALIDATORS = [
     {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator'},
     {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator'},
@@ -81,28 +78,25 @@ AUTH_PASSWORD_VALIDATORS = [
     {'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator'},
 ]
 
+# Internationalization
 LANGUAGE_CODE = 'en-us'
 TIME_ZONE = 'Asia/Manila'
 USE_I18N = True
 USE_TZ = True
 
+# Static files (CSS, JavaScript, Images)
 STATIC_URL = '/static/'
-STATICFILES_DIRS = [BASE_DIR / "smart_trav_plan" / "static"]
+STATICFILES_DIRS = [BASE_DIR / "SmartTrav" / "static"]
 
+# Default primary key field type
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 # Authentication redirects
-#LOGIN_URL = '/accounts/login/'
-#LOGIN_REDIRECT_URL = '/accounts/dashboard/'
-#LOGOUT_REDIRECT_URL = '/accounts/login/'
+LOGIN_URL = '/smart_trav_plan/login/'
+LOGIN_REDIRECT_URL = '/smart_trav_plan/dashboard/'
+LOGOUT_REDIRECT_URL = '/smart_trav_plan/login/'
 
-LOGIN_URL = '/accounts/login/'
-LOGIN_REDIRECT_URL = '/accounts/dashboard/'
-LOGOUT_REDIRECT_URL = '/accounts/login/'
-LOGOUT_REDIRECT_URL = '/accounts/login/'
-
-
-from django.contrib.messages import constants as messages
+# Message tags
 MESSAGE_TAGS = {
     messages.ERROR: 'danger',
 }
