+++ smart_trav_plan/smart_trav_plan/urls.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+from django.urls import path, include
+from django.views.generic import RedirectView
+
+urlpatterns = [
+    path('admin/', admin.site.urls),
+    path('', RedirectView.as_view(url='/login/', permanent=False)),  # ✅ Redirect root to login
+    path('', include('SmartTrav.urls')),
+]
